fix(importer): Sage r3 #108 — keep handle internal, fix collision source path, dedupe outcome + fixtures
Addresses 4 findings from Sage R3 (offline, dispatcher unreachable):

  R3 #1 (Security, blocker) — pai-pack-importer.ts handle apply.
  `importPaiPackFromPlan` + `planPaiPackImportHandle` +
  `PaiPackImportPlanHandle` are no longer re-exported from
  `src/index.ts`. The opaque handle relies on `castHandle` to recover
  the internal plan; with the trio on the public barrel, a JS caller
  could forge a handle object with attacker-controlled `routedFiles`
  whose targets escape `somaHome`, and the cached-plan apply would
  write them. Migration is the sole consumer and imports the trio
  directly from `./pai-pack-importer`; the cast is unforgeable from
  outside the module file.

  R3 #2 (Architecture, important) — same fix as #1. Internal
  optimization should not be irreversible SDK surface.

  R3 #3 (CodeQuality, suggestion) — pai-pack-importer.ts flat-vs-
  nested collision refusal source path. Previously built as
  `src/${packSlug}/SKILL.md` using the kebab-cased pack slug; for
  raw nested name `PAIUpgrade` colliding with pack slug
  `pai-upgrade`, the refusal would report `src/pai-upgrade/SKILL.md`,
  a path that does not exist on disk. Resolved via the matching
  `nestedIndex.raw` so the message pins the real filesystem path.

  R3 #4 (Maintainability, suggestions) — single-source for the
  cross-pack `refused-name-collision` outcome via
  `crossPackCollisionOutcome(paiPackDir, slug)`. Both bulk paths
  (apply + plan-only) emit identical rows; the reason text and
  outcome kind can no longer drift across them. Also extracted
  pack-writer fixtures (`writeFlatPack`, `writeNestedSkill`,
  `writeNestedPackShell`, `writeFlatNestedPack`) into
  `test/fixtures/pai-pack-fixtures.ts` — issue-105 and issue-108-r2
  test files now import from one canonical source.

R3 surface: 1 blocker, 1 important, 2 suggestion (strictly smaller
than R2's 2 important + 2 suggestion; the new blocker exists only
because R2 introduced the handle API — addressing it by demoting
the handle to module-internal closes that surface as well).

Tests stay green: 686 pass, 0 fail, 2190 expect() calls.
